+++ src/main/java/com/advantech/facade/TestLineTypeFacade.java
@@ -59,7 +59,7 @@ public class TestLineTypeFacade extends BasicLineTypeFacade {
     private WebServiceRV rv;
 
     private final int TEST_USER_NOT_IN_SYSTEM_SIGN = -1, TEST_USER_NOT_IN_XML_SIGN = 2;
-    
+
     private Double testSaltProductivity;
 
     @PostConstruct
@@ -72,13 +72,13 @@ public class TestLineTypeFacade extends BasicLineTypeFacade {
         PEOPLE_NOT_MATCH = new HashMap();
         testSaltProductivity = p.getTestSaltProductivity();
         this.initMap();
-        
+
         //Re init the alarm sign again when line name changed or sensor name changed
 //        this.initAlarmSign();
     }
 
     @Override
-    protected void initMap() {
+    public void initMap() {
         super.dataMap.clear();
         PEOPLE_NOT_MATCH.clear();
 
@@ -99,7 +99,7 @@ public class TestLineTypeFacade extends BasicLineTypeFacade {
 
             processingJsonObject = new JSONObject();
             boolean isInTheWebService = false;
-            
+
             //In byPass hours, system show normal sign 
             //no matter user's productivity is under the standard on browser.
             boolean byPassFlag = isByPassCurrentHours();
@@ -109,7 +109,7 @@ public class TestLineTypeFacade extends BasicLineTypeFacade {
                 String jobnumber = record.getUserId();
                 String userName = record.getUserName();
                 Double productivity = record.getProductivity();
-                
+
                 //※ Productivity changed here ※
                 productivity = productivity + testSaltProductivity;
 
@@ -176,7 +176,9 @@ public class TestLineTypeFacade extends BasicLineTypeFacade {
     @Override
     public void initAlarmSign() {
         List l = almService.findAll();
-        almService.delete(l);
+        if (!l.isEmpty()) {
+            almService.delete(l);
+        }
         almService.insert(this.mapToAlarmSign(dataMap));
     }
